feat(client): add loop_forever method
This commit introduces a new method `loop_forever` in the `Client` class. This allows for an infinite loop that keeps the client running until a KeyboardInterrupt is detected, upon which it logs a shutdown message.

Changes:
- Added a new attribute `__http_server_thread` to track the http server thread
- Modified the initialization of the http server thread to assign it to `__http_server_thread`
- Created a new method `loop_forever` which checks if the http server thread is running before entering an infinite loop
- An exception for KeyboardInterrupt is implemented to handle graceful shutdown of the client.

This change makes it easier to manage the lifecycle of the client object in code, especially in scenarios when an application needs to maintain a persistent connection. Previously, there was no built-in way to keep the client running indefinitely.