refactor(client): split subscribe_topic into two methods
The `subscribe_topic` function in the `Client` class has been refactored. Previously it was capable of subscribing to either a single topic or a list of topics. This was potentially confusing and could lead to misuse.

Changes:

- The function has been split into two: `subscribe_topic` and `subscribe_topics`. The former now only accepts a single topic as a string, while the latter accepts a list of topics.
- The logic for subscribing to topics has been moved to a private function `_subscribe_topics`.
- The documentation for both functions has been updated to reflect these changes.

This change will make the API clearer and easier to use, as it is now explicit whether a single topic or multiple topics are being subscribed to.