feat(client): add await_response_on_send_event method
In this change, a new method `await_response_on_send_event` is added to the Client class in the `event_connector_lib` module. This method sends an event and waits for a response within a specified timeout. If the response is not received within the timeout period, an `AwaitEventResponseTimeout` exception is raised.

Changes:
- A new TimeoutError subclass, `AwaitEventResponseTimeout`, is introduced.
- The new method `await_response_on_send_event` is added to the Client class. This method sends an event to the broker and then waits up to the specified number of seconds for a response.
- An example usage of the new method is also included in the method's docstring.