refactor(event_connector_lib): Move ModuleType enum to utils module
In this commit, the ModuleType enum, which enumerates different types of ZKMS Modules, is moved from the `client.py` file to the `utils.py` file in the `event_connector_lib` directory.

Changes:
- Removed import of `enum` and the entire `ModuleType` class definition from `client.py`
- Added import of `enum` and the `ModuleType` class definition to `utils.py`

This change makes the `ModuleType` enum more accessible to other modules and components that may need to use it, without having to import the entire `client.py` module.