feat(event): add get_payload_object function
A new method 'get_payload_object' is added to the Event class in utils.py. This function allows for the retrieval of an object from the event payload using a JSON path.

Changes:
- Added a new method 'get_payload_object' in Event class
- This method takes a string argument 'json_path' representing JSON path to the desired object
- Returns the object retrieved from the JSON path or None if it doesn't exist
- Raises a KeyError if any of the keys in the JSON path do not exist

This new feature will provide a more flexible way to access data within the event payload, especially for nested objects.